pyproject.toml
[tool.poetry]
name = "pizza-delivery-api"
version = "0.1.0"
description = "Projeto em fastapi para delivery de pizza"
authors = ["Soilenyis <soilenyis@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
sqlalchemy = "^2.0.25"
psycopg2-binary = "^2.9.9"
sqlalchemy-utils = "^0.41.1"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
alembic = "^1.13.1"
email-validator = "^2.1.0.post1"
werkzeug = "^3.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
